# Improve the computation of the auxiliary variables in AuxIVA and ILRMA