@jazcarretao 2020-10-27
Good spot and thanks for the profiling! Have you run the tests just in case?