@mori97 2020-10-27
I ran the tests with `nosetests`.
Also, I have confirmed the outputs of `auxiva` and `ilrma` do not change after this modification by `numpy.all_close`.